+++ programs/drift/src/error.rs
@@ -523,6 +523,10 @@ pub enum ErrorCode {
     CantReclaimRent,
     #[msg("InsuranceFundOperationPaused")]
     InsuranceFundOperationPaused,
+    #[msg("NoUnsettledPnl")]
+    NoUnsettledPnl,
+    #[msg("PnlPoolCantSettleUser")]
+    PnlPoolCantSettleUser,
 }
 
 #[macro_export]